Move system_call_id / thunk_id to separate source files
