+++ types/chain.hpp
@@ -1,78 +1,5 @@
 namespace koinos { namespace chain {
 
-// Use generate_ids.py to generate the system call id
-enum class system_call_id : uint32
-{
-   prints = 0x9b229941,
-   apply_block = 0x94ab4ff5,
-   apply_transaction = 0x9d8b55da,
-   apply_reserved_operation = 0x9aa85924,
-   apply_upload_contract_operation = 0x9e6ea937,
-   apply_execute_contract_operation = 0x92184686,
-   apply_set_system_call_operation = 0x9579a45c,
-   db_put_object = 0x971ec7a2,
-   db_get_object = 0x9766a8fb,
-   db_get_next_object = 0x99a398e8,
-   db_get_prev_object = 0x9bd3767c,
-   execute_contract = 0x98c12cfe,
-   get_entry_point = 0x9ff207dd,
-   get_contract_args_size = 0x9b0d8fd9,
-   get_contract_args = 0x9fbba198,
-   set_contract_return = 0x9f49cdea,
-   exit_contract = 0x98df75b0,
-   get_head_info = 0x956fb22d,
-   hash = 0x99770e04,
-   recover_public_key = 0x95724a88,
-   verify_block_signature = 0x9d1c3c89,
-   verify_merkle_root = 0x996e24b9,
-   get_transaction_payer = 0x9db35086,
-   get_max_account_resources = 0x90f14f8d,
-   get_transaction_resource_limit = 0x9940f685,
-   get_last_irreversible_block = 0x953d2e37,
-   get_caller = 0x94176c5f,
-   require_authority = 0x9491e528,
-   get_transaction_signature = 0x9dc3ec34,
-   get_contract_id = 0x95e30f50,
-   get_head_block_time = 0x941d5ab9
-};
-
-// Use generate_ids.py to generate the thunk id
-enum class thunk_id : uint32
-{
-   prints = 0x8f6df54d,
-   verify_block_header = 0x8d425aac,
-   apply_block = 0x8d6d31a8,
-   apply_transaction = 0x8981b0df,
-   apply_reserved_operation = 0x8b3c14f6,
-   apply_upload_contract_operation = 0x8882a55e,
-   apply_execute_contract_operation = 0x85e882eb,
-   apply_set_system_call_operation = 0x86f92c8c,
-   db_put_object = 0x82038de5,
-   db_get_object = 0x8862a0d8,
-   db_get_next_object = 0x86e45047,
-   db_get_prev_object = 0x8d57e8fd,
-   execute_contract = 0x8a43fe83,
-   get_entry_point = 0x807eb765,
-   get_contract_args_size = 0x83378e86,
-   get_contract_args = 0x8e189d86,
-   set_contract_return = 0x86b86275,
-   exit_contract = 0x81f61f9f,
-   get_head_info = 0x89df34c4,
-   hash = 0x8aaaf547,
-   recover_public_key = 0x804f5450,
-   verify_block_signature = 0x8fb9a59b,
-   verify_merkle_root = 0x8ed9ddcb,
-   get_transaction_payer = 0x86a87bf5,
-   get_max_account_resources = 0x842c6c81,
-   get_transaction_resource_limit = 0x8bdf81a1,
-   get_last_irreversible_block = 0x80c3b893,
-   get_caller = 0x82312501,
-   require_authority = 0x8a06717d,
-   get_transaction_signature = 0x83441b23,
-   get_contract_id = 0x8df4ad21,
-   get_head_block_time = 0x8e76d5b4
-};
-
 enum class privilege : uint8
 {
    kernel_mode,
